wiring up new auth to controllers and UI

+++ server/controllers/package.js
@@ -4,16 +4,17 @@ const upload = multer({
   dest: 'uploads/' 
 });
 const model = require('../models/PackageModel');
+const middleware = require('./middleware/auth');
 const utils = require('./utils');
 const AppError = require('../lib/AppError');
 
 router.post('/', async (req, res) => {
-  let package = req.body;
-  
-  // TODO
-  // package.owner = req.user.username;
-  package.owner = 'bob';
-  package.organization = 'myorg';
+  let pkg = req.body || {};
+
+  pkg.owner = req.session.username;
+
+  let writeAccess = await middleware.canWritePackage(pkg, req, res);
+  if( !writeAccess ) return;
 
   try {
     package = await model.create(package);
@@ -30,6 +31,9 @@ router.get('/:name', async(req, res) => {
     return utils.handleError(res, e);
   }
 
+  let readAccess = await middleware.canReadPackage(packageName, req, res);
+  if( !readAccess ) return;
+
   try {
     let package = await model.get(packageName);
     res.json(package);
@@ -38,11 +42,20 @@ router.get('/:name', async(req, res) => {
   }
 });
 
-router.patch('/', async(req, res) => {
-  let package = req.body;
+router.patch('/:name', async(req, res) => {
+  let packageName = req.params.name;
+  if( !packageName ) {
+    let e = new AppError('Package name or id required. /api/package/:name', AppError.ERROR_CODES.MISSING_ATTRIBUTE)
+    return utils.handleError(res, e);
+  }
+
+  let update = req.body;
 
   try {
-    package = await model.update(package);
+    let writeAccess = await middleware.canWritePackage(packageName, req, res);
+    if( !writeAccess ) return;
+
+    package = await model.update(update);
     res.json(package);
   } catch(e) {
     utils.handleError(res, e);
@@ -57,6 +70,9 @@ router.delete('/:name', async (req, res) => {
   }
 
   try {
+    let writeAccess = await middleware.canWritePackage(packageName, req, res);
+    if( !writeAccess ) return;
+  
     await model.delete(packageName);
     res.status(204);
   } catch(e) {
@@ -67,6 +83,9 @@ router.delete('/:name', async (req, res) => {
 router.post('/:name/createRelease', async (req, res) => {
   let releaseInfo = req.body;
   let packageName = req.params.name;
+
+  let writeAccess = await middleware.canWritePackage(packageName, req, res);
+  if( !writeAccess ) return;
   
   try {
     package = await model.createRelease(packageName, releaseInfo);
@@ -76,7 +95,12 @@ router.post('/:name/createRelease', async (req, res) => {
   }
 });
 
-router.post('/addFile', upload.any(), async (req, res) => {
+router.post('/:name/addFile', upload.any(), async (req, res) => {
+  let packageName = req.params.name;
+
+  let writeAccess = await middleware.canWritePackage(packageName, req, res);
+  if( !writeAccess ) return;
+
   // req.files is array of `photos` files 
   // req.body will contain the text fields, if there were any
